+++ docs/src/changelog.md
@@ -7,7 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - ![][badge-feature] Added fuzzy c-means
 - ![][badge-enhancement] added support Lukasiewicz, drastic, nilpotent and Hamacher T-norms and corresponding S-norms.
 - ![][badge-enhancement] dont build anonymous functions during mamdani inference, but evaluate output directly. Now defuzzifiers don't take a function as input, but an array.
-
+- ![][badge-feature] added piecewise linear membership functions
+![][badge-feature] added parser for Fuzzy Control Language 
 ## v0.1.0 -- 2023-01-10
 
 [view release on GitHub](https://github.com/lucaferranti/FuzzyLogic.jl/releases/tag/v0.1.0)